jsl/requirements.txt
Django==1.11.4
django-bootstrap-form==3.3
pytz==2017.2
djangorestframework==3.6.4
factory-boy==2.9.2

